s04e02/requirements.txt
python-dotenv
requests
deepagents
langchain-openai
